Split syscheck_audit.c
Created new files to hold functions with specific purpouses when handling audit in the context of FIM.
- audit_parse.c: Holds all functions responsible for parsing audit messages into FIM events.
- audit_rule_handling.c: Holds all functions responsible for loading and removing FIM whodata rules from the audit kernel module.
- audit_healthcheck: Holds all functions responsible for checking auditd is working properly on start up.
